FittingFunction

Documentation for FittingFunction.

A set of functions I am often using in carrying out regressions.

FittingFunction.Band

FittingFunction.CPL

FittingFunction.Counts2Mag

FittingFunction.Extinction

FittingFunction.GaussAbs

FittingFunction.Gaussian

FittingFunction.Mag2Counts

FittingFunction.PL

FittingFunction.Pol2Stokes

FittingFunction.SBPL

FittingFunction.SBPL2

FittingFunction.Stokes2Pol

FittingFunction.TauVoigt

FittingFunction.VoigtFunctTG

FittingFunction.XAbs

FittingFunction.Band — Method

Band(E,α,β,E₀,A)

Compute the so-called 'Band function', at input energy 'E'. 'α' is the low-energy index, 'β' the high-energy index and 'E₀' the break energy. 'A' is the normalization. The peak-energy is 'Ep = (2+α)E₀'.

Examples

Band(300.,-2.,-1.,200.,1.)

# output

-0.45304697140984085

FittingFunction.CPL — Method

CPL(E,N,α,Ec;E0=1.)

Computes a cut-off power-law with spectral index 'α' and normalization 'N' at input 'E' and cut-off energy 'Ec'. You can normalize the power-law at 'E0'.

Examples

CPL(3.,1.,-1.,2.)

# output

0.0743767200494766

FittingFunction.Counts2Mag — Method

Counts2Mag(cts,ects;zp=25.0,ezp=0.0)

Convert counts (or flux) 'cts', with uncertainty 'ects', to magnitude scale with zero-point 'zp' and relative uncertainty.

Examples

Counts2Mag(100,10)

# output

(20.0, 0.10857362047581294)

FittingFunction.Extinction — Method

Extinction(wave,EBV;gal="SMC",Rv=FFGals["SMC"],z=0.)

Compute the UV/optical/extinction at the input wavelength 'wave', in Angstrom. 'EBV' is E(B-V) in magnitudes, 'gal' is one of the galaxy extinction recipes listed in the 'FFGals' (exported) dictionary. 'Rv' is the seelctive extinction and 'z' is the redshift of the absorpber. References about the adopted extinction curves are discussed in the documentation of the DustExtinction package.

Examples

Extinction(5500.,1,gal="MW",Rv=FFGals["MW"],z=0.)

# output

0.06017150537641227

FittingFunction.GaussAbs — Method

GaussAbs(E,Ed,σ,El)

Computes a Gaussian absorption with depth 'Ed' at energy 'E' with center in 'El' and width (FWHM/2.35) 'σ'. 'A.

Examples

GaussAbs(2.,1.,0.1,2.)

# output

0.018510395162776326

FittingFunction.Gaussian — Method

Gaussian(E,A,σ,El)

Computes a Gaussian at energy 'E' with center in 'El' and width (FWHM/2.35) 'σ'. 'A is the normalization.

Examples

Gaussian(3.,1.,0.1,2.)

# output

4.009419869036418

FittingFunction.Mag2Counts — Method

Mag2Counts(mag,emag;zp=25.0)

Convert magnitudes 'mag' with uncertainty 'emag' to counts (or flux) with zero-point 'zp'.

Examples

Mag2Counts(20,0.1)

# output

(100.0, 9.210340371976184)

FittingFunction.PL — Method

PL(E,N,α;E0=1.)

Computes a power-law with spectral index 'α' and normalization 'N' at input 'E'. You can normalize the power-law at 'E0'.

Examples

PL(3.,1.,-1.)

# output

0.3333333333333333

FittingFunction.Pol2Stokes — Method

Pol2Stokes(i, p, t, c; ep=0.0, et=0.0, ec=0.0)

Computes the Stokes parameters for a given polarisation degree and position angle. 'i' is total intensity, 'p' the polarisation degree, 't' the position angle (randians) and 'c' the circular polarisation. The errors 'ep', 'ew, and 'e' are optional. The ouput is: intensity 'q', 'u', 'v' Stokes parameters and respective errors.

Examples

Pol2Stokes(1.,0.1,0.1,0.)

# output

(1.0, 0.09800665778412417, 0.019866933079506124, 0.0, 0.0, 0.0, 0.0)

FittingFunction.SBPL — Method

SBPL(E,N,α,β,Eb)

Computes a smoothly joint broken power-law with spectral index 'α' and 'β', a break at 'Eb', normalization 'N' and input energy'E'.

Examples

SBPL(6.,1.,-1.,-1.5,5.)

# output

0.15214515486254615

FittingFunction.SBPL2 — Method

SBPL2(E,N,α,β,γ,Eb1,Eb2)

Computes a smoothly joint double broken power-law with spectral index 'α', 'β' and 'γ', and two breaks at 'Eb1' and 'Eb2'. The normalization is 'N' and the input energy 'E'.

Examples

SBPL2([4.,6.,8.],1.,-1.,-1.5,-2.,5.,7.)

# output

3-element Vector{Float64}:
 0.25
 0.06804138174397717
 0.04133986423538423

FittingFunction.Stokes2Pol — Method

Stokes2Pol(i, q, u, v; eq=0.0, eu=0.0, ev=0.0)

Computes the polarisation degree and position angle given the Stokes parameters: 'i', total intensity, and 'q', 'u' and 'v'. The errors 'eq', 'eu, and 'ev' are optional. The ouput is: intensity 'i', polarisation degree, 'p', position angle (randians) 'theta' and circular polarisation 'chi', with respective errors.

Examples

Stokes2Pol(1.,0.1,0.1,0.)

# output

(1.0, 0.14142135623730953, 0.39269908169872414, 0.0, 0.0, 0.0, 0.0)

FittingFunction.TauVoigt — Method

TauVoigt(λ,NHI,DopplerBroadening,z,transition)

Computes the opacity due to a given absorption transition. $\lambda$ are the wavelenghths ($cm$), $N$ the column density ($cm^{-2}$), 'DopplerBroadening' is expressed in ($cm~s^{-1}$), 'z' is the source redshift and 'transition' is a tuple formed by the central wawelength ($cm$), the oscillator strength and the damping coefficient ($s^{-1}$) for the given transition.

Examples

TauVoigt(range(start=1494.5,stop=1495.5,step=0.1) .* 1e-8,1e18,1e5,0.,[1495.05*1e-8,0.54,8.106e8])

# output

11-element Vector{Float64}:
  0.5408331361420329
  0.8079604230648749
  1.3357686994197793
  2.618871809580839
  7.282358478732964
 66.40911467182524
 66.40911467182524
  7.282358478732964
  2.618871809580839
  1.3357686994184856
  0.8079604230648749

FittingFunction.VoigtFunctTG — Method

VoigtFuncTG(a,v)

Approximates the Voigt (or line broadening function) function. It was introduced by Tepper-Garcia (2006). It is of high accuracy provided that $a \le 10^{-4}$. The $a$ and $v$ parameters are defined as: $a = \frac{\Gamma \lambda_c}{4\pi b 10^{13}}$ and $v = \frac{(\lambda_c - \lambda)c}{b \lambda_c \sqrt{2 \log 2}}$, where $\lambda_c$ is the central wavelength of the transition, $c$ the speed of light and $b$ the Doppler broadening factor.

Examples

FittingFunction.VoigtFunctTG(1e-5,1.)

# output

0.36788094737611143

FittingFunction.XAbs — Method

XAbs(E; NH=1e20, z=0)

Computes the effective absorption cross section per hydrogen atom following Morrison & McCammon (1983). 'E' is the energy (KeV, in the 0.03-10 KeV range), 'NH' is the column density of hydrogen atom (particle per square cm) and 'z' is the redshift.

Examples

XAbs([0.5,1.25,2.], NH=1e20, z=0)

# output

3-element Vector{Float64}:
 0.9290803992951834
 0.9868927382232576
 0.9957079871273042