FittingFunction

Documentation for FittingFunction.

A set of functions I am often using in carrying out regressions.

FittingFunction.Band

FittingFunction.CPL

FittingFunction.Counts2Mag

FittingFunction.Extinction

FittingFunction.FindRebinSchema

FittingFunction.GaussAbs

FittingFunction.Gaussian

FittingFunction.Mag2Counts

FittingFunction.PL

FittingFunction.SBPL

FittingFunction.SBPL2

FittingFunction.XAbs

FittingFunction.Band — Method

Band(E,α,β,E₀,A)

Compute the so-called 'Band function', at input energy 'E'. 'α' is the low-energy index, 'β' the high-energy index and 'E₀' the break energy. 'A' is the normalization. The peak-energy is 'Ep = (2+α)E₀'.

Examples

Band(300.,-2.,-1.,200.,1.)

# output

-0.45304697140984085

FittingFunction.CPL — Method

CPL(E,N,α,Ec;E0=1.)

Computes a cut-off power-law with spectral index 'α' and normalization 'N' at input 'E' and cut-off energy 'Ec'. You can normalize the power-law at 'E0'.

Examples

CPL(3.,1.,-1.,2.)

# output

0.0743767200494766

FittingFunction.Counts2Mag — Method

Counts2Mag(cts,ects;zp=25.0,ezp=0.0)

Convert counts (or flux) 'cts', with uncertainty 'ects', to magnitude scale with zero-point 'zp' and relative uncertainty.

Examples

Counts2Mag(100,10)

# output

(20.0, 0.10857362047581294)

FittingFunction.Extinction — Method

Extinction(wave,EBV;gal="SMC",Rv=FFGals["SMC"],z=0.)

Compute the UV/optical/extinction at the input wavelength 'wave', in Angstrom. 'EBV' is E(B-V) in magnitudes, 'gal' is one of the galaxy extinction recipes listed in the 'FFGals' (exported) dictionary. 'Rv' is the seelctive extinction and 'z' is the redshift of the absorpber. References about the adopted extinction curves are discussed in the documentation of the DustExtinction package.

Examples

Extinction(5500.,1,gal="MW",Rv=FFGals["MW"],z=0.)

# output

0.06017150537641227

FittingFunction.FindRebinSchema — Method

FindRebinSchema(x,ey;minSN=5)::AbstractVector{Real}

Compute the rebin schema to guarantee that the S/N is at least 'minSN' in each bin (or channel). 'x' and 'ex' are the input data and relative uncertainties.

Examples

x = [1.,2.,3.,4.,]
ex = [0.1,0.5,0.6,0.05]

FindRebinSchema(x,ex)

# output

3-element Vector{Real}:
 1
 3
 4

FittingFunction.GaussAbs — Method

GaussAbs(E,Ed,σ,El)

Computes a Gaussian absorption with depth 'Ed' at energy 'E' with center in 'El' and width (FWHM/2.35) 'σ'. 'A.

Examples

GaussAbs(2.,1.,0.1,2.)

# output

0.018510395162776326

FittingFunction.Gaussian — Method

Gaussian(E,A,σ,El)

Computes a Gaussian at energy 'E' with center in 'El' and width (FWHM/2.35) 'σ'. 'A is the normalization.

Examples

Gaussian(3.,1.,0.1,2.)

# output

4.009419869036418

FittingFunction.Mag2Counts — Method

Mag2Counts(mag,emag;zp=25.0)

Convert magnitudes 'mag' with uncertainty 'emag' to counts (or flux) with zero-point 'zp'.

Examples

Mag2Counts(20,0.1)

# output

(100.0, 9.210340371976184)

FittingFunction.PL — Method

PL(E,N,α;E0=1.)

Computes a power-law with spectral index 'α' and normalization 'N' at input 'E'. You can normalize the power-law at 'E0'.

Examples

PL(3.,1.,-1.)

# output

0.3333333333333333

FittingFunction.SBPL — Method

SBPL(E,N,α,β,Eb)

Computes a smoothly joint broken power-law with spectral index 'α' and 'β', a break at 'Eb', normalization 'N' and input energy'E'.

Examples

SBPL(6.,1.,-1.,-1.5,5.)

# output

0.15214515486254615

FittingFunction.SBPL2 — Method

SBPL2(E,N,α,β,γ,Eb1,Eb2)

Computes a smoothly joint double broken power-law with spectral index 'α', 'β' and 'γ', and two breaks at 'Eb1' and 'Eb2'. The normalization is 'N' and the input energy 'E'.

Examples

SBPL2([4.,6.,8.],1.,-1.,-1.5,-2.,5.,7.)

# output

3-element Vector{Float64}:
 0.25
 0.06804138174397717
 0.04133986423538423

FittingFunction.XAbs — Method

XAbs(E; NH=1e20, z=0)

Computes the effective absorption cross section per hydrogen atom following Morrison & McCammon (1983). 'E' is the energy (KeV, in the 0.03-10 KeV range), 'NH' is the column density of hydrogen atom (particle per square cm) and 'z' is the redshift.

Examples

XAbs([0.5,1.25,2.], NH=1e20, z=0)

# output

3-element Vector{Float64}:
 0.9290803992951834
 0.9868927382232576
 0.9957079871273042