FittingFunction

Documentation for FittingFunction.

A set of functions I am often using in carrying out regressions.

FittingFunction.Band

FittingFunction.Extinction

FittingFunction.PL

FittingFunction.SBPL

FittingFunction.XAbs

FittingFunction.Band — Method

Band(E,α,β,E₀,A)

Compute the so-called 'Band function', at input energy 'E'. 'α' is the low-energy index, 'β' the high-energy index and 'E₀' the break energy. 'A' is the normalization. The peak-energy is 'Ep = (2+α)E₀'.

Examples

Band(300.,-2.,-1.,200.,1.)

# output

-0.45304697140984085

FittingFunction.Extinction — Method

Extinction(wave,EBV;gal="SMC",Rv=FFGals["SMC"],z=0.)

Compute the UV/optical/extinction at the input wavelength 'wave', in Angstrom. 'EBV' is E(B-V) in magnitudes, 'gal' is one of the galaxy extinction recipes listed in the 'FFGals' (exported) dictionary. 'Rv' is the seelctive extinction and 'z' is the redshift of the absorpber. References about the adopted extinction curves are discussed in the documentation of the DustExtinction package.

Examples

Extinction(5500.,1,gal="MW",Rv=FFGals["MW"],z=0.)

# output

0.06017150537641227

FittingFunction.PL — Method

PL(E,N,α;E0=1.)

Computes a power-law with spectral index 'α' and normalization 'N' at input 'E'. You can normalize the power-law at 'E₀'.

Examples

PL(3.,1.,-1.)

# output

0.3333333333333333

FittingFunction.SBPL — Method

SBPL(E,N,α,β,Eb)

Computes a smoothly joint broken power-law with spectral index 'α' and 'β', a break at 'Eb' and normalization 'N' at input 'E'.

Examples

SBPL(6.,1.,-1.,-1.5,5.)

# output

0.15214515486254615

FittingFunction.XAbs — Method

XAbs(E; NH=1e20, z=0)

Computes the effective absorption cross section per hydrogen atom following Morrison & McCammon (1983). 'E' is the energy (KeV, in the 0.03-10 KeV range), 'NH' is the column density of hydrogen atom (particle per square cm) and 'z' is the redshift.

Examples

XAbs([0.5,1.25,2.], NH=1e20, z=0)

# output

3-element Vector{Float64}:
 0.9290803992951834
 0.9868927382232576
 0.9957079871273042